Rename and refactor the ODF formats
